Fix rendering bug
Not sure why it worked before, but right now render doesn't always cause
the thing to re-render. For now I just connected it to a timer so that
it renders at approximately 60 FPS. I'll need to investigate how do this
properly (i.e. make it render at the screen's referesh rate)